Backend/requirements.txt
asgiref==3.3.1
auditlog3==1.0.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.8.2
djangorestframework==3.12.2
idna==2.10
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2